Removed the need for a restart making it easier. Added verbosity.
Some steps have been removed because they are unnecessary, for example the restart. These changes make the install process a lot less involved for the user thus making it a lot easier. The only user interaction needed now is to open and close Firefox before the cshell install, and for them to accept both certificates ( https://localhost:14186/id/  and https://vpn.fct.unl.pt/) and connect.
Commented out the upgrade step because I feel like it's unnecessary, and a user might want some old version of something installed don their computer so we shouldn't update if for them.
Added verbosity to the script so the user knows whats going on while the script runs.